feat: split handle classes?
